requirements.txt
monai>=0.9
nibabel
scikit-image
matplotlib
Pillow
tensorboard
gdown
torchVision
tqdm
psutil
pandas
einops
sentence_transformers
tf-keras
sentencepiece
promptcap
transformers==4.28.1
mmf